Add basic tax, alternative minim tax and lump sum tax files
