+++ policyengine_us/parameters/gov/states/mn/tax/income/rates/single.yaml
@@ -22,7 +22,7 @@ brackets:
 
 metadata:
   type: marginal_rate
-  threshold_unit: currency_USD
+  threshold_unit: currency-USD
   rate_unit: /1
   threshold_period: year
   label: Minnesota single income tax rates
