+++ policyengine_us/variables/gov/states/mn/tax/income/mn_basic_tax.py
@@ -1,0 +1,36 @@
+from policyengine_us.model_api import *
+
+
+class mn_basic_tax(Variable):
+    value_type = float
+    entity = TaxUnit
+    label = "Minnesota basic tax calculated using tax rate schedules"
+    unit = USD
+    definition_period = YEAR
+    reference = (
+        "https://www.revenue.state.mn.us/sites/default/files/2023-01/m1_inst_21.pdf"
+        "https://www.revenue.state.mn.us/sites/default/files/2023-03/m1_inst_22.pdf"
+    )
+    defined_for = StateCode.MN
+
+    def formula(tax_unit, period, parameters):
+        filing_status = tax_unit("filing_status", period)
+        statuses = filing_status.possible_values
+        taxable_income = tax_unit("mn_taxable_income", period)
+        p = parameters(period).gov.states.mn.tax.income.rates
+        return select(
+            [
+                filing_status == statuses.SINGLE,
+                filing_status == statuses.SEPARATE,
+                filing_status == statuses.JOINT,
+                filing_status == statuses.WIDOW,
+                filing_status == statuses.HEAD_OF_HOUSEHOLD,
+            ],
+            [
+                p.single.calc(taxable_income),
+                p.separate.calc(taxable_income),
+                p.joint.calc(taxable_income),
+                p.widow.calc(taxable_income),
+                p.head_of_household.calc(taxable_income),
+            ],
+        )
